% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{glasso_manual}
\alias{glasso_manual}
\title{glasso_manual}
\usage{
glasso_manual(data, file = "./manual_selection.RData", real_network = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{file}{Path to the file, where solution is to be saved.}

\item{real_network}{Adjadency matrix of real network, if available.}
}
\description{
Creates an visualization of glasso solution.
}
\examples{
data <- huge.generator(n = 75, d = 100, graph = "scale-free", verbose = FALSE)
app <- glasso_manual(data = data$data,  real_network = data$theta)
shiny::runApp(app)
}
