% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLASSO_Visualization.R
\name{ManualSelection}
\alias{ManualSelection}
\title{ManualSelection}
\usage{
ManualSelection(data, output_file = "./rho_value.RData")
}
\arguments{
\item{data}{dataframe of observations with column-names or without.}

\item{output_file}{Name of the file, where user selected solution in stored}
}
\description{
Creates visualization tool for glasso. Allows user to choose GLASSO-solution and visualize it immediately.
}
\examples{
data <- huge.generator(n = 75, d = 100, graph = "scale-free", verbose = FALSE)
app <- ManualSelection(data = data$data)
shiny::runApp(app)
}
