% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLASSO_Visualization_with_know_network.R
\name{ManualSelection_known_network}
\alias{ManualSelection_known_network}
\title{ManualSelection_known_network}
\usage{
ManualSelection_known_network(
  data,
  output_file = "./rho_value.RData",
  real_network
)
}
\arguments{
\item{data}{Sampled data in a data.frame form.}

\item{output_file}{The name of the file if user decides to save the solution chosen.}

\item{real_network}{Real network in either matrix or sparse-matrix form}
}
\description{
Creates an visualization for GLASSO-solution path in a simulated setup.
}
\examples{
data <- huge.generator(n = 75, d = 100, graph = "scale-free", verbose = FALSE)
app <- ManualSelection_known_network(data = data$data,  real_network = data$theta)
shiny::runApp(app)
}
